added cutting scaffold and associations

+++ db/migrate/20201112205838_create_plants.rb
@@ -3,12 +3,12 @@ class CreatePlants < ActiveRecord::Migration[6.0]
     create_table :plants do |t|
       t.integer :order_id
       t.string :name
-      t.integer :unit_price
+      t.decimal :unit_price
       t.date :arrival_date
       t.string :condition
       t.date :recovery_date
       t.boolean :sold
-      t.integer :price
+      t.decimal :price
       t.integer :tag
 
       t.timestamps
